DBZ-1144 Do not translate geometry point (PostGIS) to PG point
This code was effectivelly dead. It was used only when PostGIS geometry type was created inside the public schema and was responsible for translating of PostGIS point to PG point. When we introduced the full support for PostGIS types in Debezium the code was no longer used as PostGIS types are typically created in separate schema.
So now we support

* POINT type (postgres point)
* postgis GEOMETRY in arbitrary schema (including geometry point)